+++ client/src/main/kotlin/io/github/octaviusframework/client/OctaviusClientImpl.kt
@@ -12,6 +12,7 @@ import io.github.octaviusframework.client.transaction.TransactionDefinition
 import io.github.octaviusframework.client.transaction.TransactionPropagation
 import io.github.octaviusframework.driver.session.OctaviusSessionOperations
 import io.github.octaviusframework.driver.session.TransactionIsolationLevel
+import io.github.octaviusframework.serializer.octaviusJson
 import kotlinx.serialization.json.Json
 import kotlin.time.Duration
 
@@ -26,7 +27,7 @@ import kotlin.time.Duration
 internal class OctaviusClientImpl(
     private val provider: SessionProvider,
     private val onClose: (() -> Unit)? = null,
-    dynamicJson: Json = Json,
+    dynamicJson: Json = octaviusJson,
     dynamicWriteStrategy: DynamicWriteStrategy = DynamicWriteStrategy.AUTOMATIC_WHEN_UNAMBIGUOUS
 ) : OctaviusClient {